requirements.txt
python-telegram-bot[job-queue]==21.10
python-dotenv==1.0.1
pymongo==4.11.1
requests==2.32.3

pytest
pytest-asyncio
pytest-mock
mongomock
requests-mock
coverage